Move contexts to their own folder
