+++ components/modals/index.ts
@@ -2,7 +2,4 @@ export type * from '@/components/modals/types'
 
 export { Modal } from '@/components/modals/Modal'
 export { ModalContainer } from '@/components/modals/ModalContainer'
-export {
-  ModalContext,
-  ModalContextWrapper,
-} from '@/components/modals/ModalContext'
+export { ModalContext, ModalContextWrapper } from '@/contexts/ModalContext'
